:construction: Basic rework of my travel request wizard with step identifiers.
Not currently working, but most concepts are in place.
